Allow a single clone of a repo to be reused
Set the repo and tag information in a global property
when cloning a gitexternal repo, then reuse this if another
request comes in to clone the same repo. Place the cloned location
in a variable so the caller can get the location of the location of
the repo, either cloned or referenced from previously.